requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
aiosmtplib==2.0.2
aiosqlite==0.17.0
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
Babel==2.14.0
bcrypt==4.1.2
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-admin==1.0.4
fastapi-filter==1.1.0
fastapi-mail==1.4.1
fastapi_auth_middleware==1.0.2
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
iso8601==1.1.0
Jinja2==3.1.3
kombu==5.3.6
Mako==1.3.2
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
pathlib==1.0.1
pendulum==3.0.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pypika-tortoise==0.1.6
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqladmin==0.16.1
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.37.2
text-unidecode==1.3
time-machine==2.14.1
tomli==2.0.1
tornado==6.4
tortoise-orm==0.20.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
WTForms==3.1.2
